requirements.txt
# Data pipeline - Olist E-Commerce (Google BigQuery)
# Python 3.10+

# BigQuery
google-cloud-bigquery>=3.14.0
pyarrow>=14.0.0
pandas-gbq>=0.19.0

# GCS
google-cloud-storage>=2.15.0

# Ingestion & ETL
pandas>=2.0.0

# Data quality
great-expectations>=0.18.0

# Kaggle download (optional)
kaggle>=1.6.0

# Analysis
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
db-dtypes>=1.2.0

# Orchestration (optional)
prefect>=2.14.0
dagster>=1.7.0
dagster-webserver>=1.7.0

# Meltano CLI (for tap-csv -> target-bigquery pipeline)
meltano>=3.0.0

# Environment (optional)
python-dotenv>=1.0.0
